refactor all east-asia-related code into seperated files
