+++ skyfield/tests/test_almanac.py
@@ -1,3 +1,4 @@
+import skyfield.almanac_east_asia
 from skyfield import api, almanac
 
 half_minute = 1.0 / 24.0 / 60.0 / 2.0
@@ -103,45 +104,45 @@ def test_close_start_and_end():
 def test_solar_terms():
     ts = api.load.timescale()
     e = api.load('de421.bsp')
-    
+
     # https://en.wikipedia.org/wiki/Lichun
-    
+
     t0 = ts.utc(2019, 2, 2)
     t1 = ts.utc(2019, 2, 5)
-    t, y = almanac.find_discrete(t0, t1, almanac.solar_terms(e))
+    t, y = almanac.find_discrete(t0, t1, skyfield.almanac_east_asia.solar_terms(e))
     t.tt += half_minute
     strings = t.utc_strftime('%Y-%m-%d %H:%M')
     print(strings)
     assert strings == ['2019-02-04 03:14']
     assert (y == (21)).all()
-    
+
     # https://en.wikipedia.org/wiki/Lixia
-    
+
     t0 = ts.utc(2019, 5, 4)
     t1 = ts.utc(2019, 5, 6)
-    t, y = almanac.find_discrete(t0, t1, almanac.solar_terms(e))
+    t, y = almanac.find_discrete(t0, t1, skyfield.almanac_east_asia.solar_terms(e))
     t.tt += half_minute
     strings = t.utc_strftime('%Y-%m-%d %H:%M')
     print(strings)
     assert strings == ['2019-05-05 19:03']
     assert (y == (3)).all()
-    
+
     # https://en.wikipedia.org/wiki/Liqiu
-    
+
     t0 = ts.utc(2019, 8, 6)
     t1 = ts.utc(2019, 8, 8)
-    t, y = almanac.find_discrete(t0, t1, almanac.solar_terms(e))
+    t, y = almanac.find_discrete(t0, t1, skyfield.almanac_east_asia.solar_terms(e))
     t.tt += half_minute
     strings = t.utc_strftime('%Y-%m-%d %H:%M')
     print(strings)
     assert strings == ['2019-08-07 19:13']
     assert (y == (9)).all()
-    
+
     # https://en.wikipedia.org/wiki/Lidong
-    
+
     t0 = ts.utc(2018, 11, 7)
     t1 = ts.utc(2018, 11, 9)
-    t, y = almanac.find_discrete(t0, t1, almanac.solar_terms(e))
+    t, y = almanac.find_discrete(t0, t1, skyfield.almanac_east_asia.solar_terms(e))
     t.tt += half_minute
     strings = t.utc_strftime('%Y-%m-%d %H:%M')
     print(strings)
